requirements.txt
Bottleneck @ file:///C:/ci_310/bottleneck_1642078621231/work
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
greenlet==1.1.2
idna==3.3
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
numexpr @ file:///C:/ci_310/numexpr_1642084889105/work
numpy @ file:///C:/ci_310/numpy_and_numpy_base_1643798589088/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1645002519186/work
psycopg2==2.9.3
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==1.4.31
urllib3==1.26.8
wincertstore==0.2
